requirements.txt
openai
streamlit
numpy
pandas
plotly
python-dotenv
langchain
langchain_openai
langchain_community
langchain_core
langgraph
langgraph-checkpoint
PyPDF2
pdfplumber
scikit-learn
pydantic>=2.0.0
networkx
typing-extensions
